chore: move image and detailer nodes to canonical package
