+++ tests/test_registry_scanner_safety.py
@@ -22,13 +22,16 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/common/serialization.py",
     "easyuse_anima/common/values.py",
     "easyuse_anima/image/__init__.py",
+    "easyuse_anima/image/detailer.py",
     "easyuse_anima/image/geometry.py",
+    "easyuse_anima/image/scaling.py",
     "easyuse_anima/infrastructure/__init__.py",
     "easyuse_anima/infrastructure/comfy/__init__.py",
     "easyuse_anima/infrastructure/comfy/capabilities.py",
     "easyuse_anima/infrastructure/comfy/invocation.py",
     "easyuse_anima/infrastructure/comfy/resources.py",
     "easyuse_anima/nodes/__init__.py",
+    "easyuse_anima/nodes/image_nodes.py",
     "easyuse_anima/prompt/__init__.py",
 }
 STATIC_IMPORT_FROM_RE = re.compile(
@@ -127,6 +130,9 @@ class RegistryScannerSafetyTests(unittest.TestCase):
             "easyuse_anima/infrastructure/comfy/capabilities.py",
             "easyuse_anima/infrastructure/comfy/invocation.py",
             "easyuse_anima/infrastructure/comfy/resources.py",
+            "easyuse_anima/image/detailer.py",
+            "easyuse_anima/image/scaling.py",
+            "easyuse_anima/nodes/image_nodes.py",
             "nodes.py",
             "prompt_translation.py",
             "settings.py",
